Remove blob store references for current Ruby builds
because they are not maintained and point to old builds.  The buildpack
will automatically fall back to loading the builds from the Heroku bucket
on S3, e.g. https://s3.amazonaws.com/heroku-buildpack-ruby/ruby-1.9.3.tgz

Leaving one old build in there because the code assumes that 'blobs' exists
and is an array.

Ref http://bugs.activestate.com/show_bug.cgi?id=102243